Move and restructure code.
